+++ .gitattributes
@@ -1,5 +1,6 @@
 src/openapi/**/* linguist-generated=true
 openapi.json linguist-generated=true
 *.woff2 linguist-vendored=true
+*.vvproj linguist-generated=true
 
 * text=auto eol=lf
