+++ src/store/project.ts
@@ -1,5 +1,3 @@
-import semver from "semver";
-import { z } from "zod";
 import { getBaseName } from "./utility";
 import { createPartialStore, Dispatch } from "./vuex";
 import { generateSingingStoreInitialScore } from "./singing";
@@ -9,75 +7,14 @@ import {
   AudioItem,
   ProjectStoreState,
   ProjectStoreTypes,
-  tempoSchema,
-  timeSignatureSchema,
-  trackSchema,
 } from "@/store/type";
 
-import { AccentPhrase } from "@/openapi";
-import {
-  audioKeySchema,
-  EngineId,
-  engineIdSchema,
-  presetKeySchema,
-  speakerIdSchema,
-  styleIdSchema,
-} from "@/type/preload";
 import { getValueOrThrow, ResultError } from "@/type/result";
-import {
-  DEFAULT_BEAT_TYPE,
-  DEFAULT_BEATS,
-  DEFAULT_BPM,
-  DEFAULT_TPQN,
-} from "@/sing/storeHelper";
-
-const DEFAULT_SAMPLING_RATE = 24000;
+import { LatestProjectType } from "@/domain/project/schema";
+import { migrateProjectFileObject } from "@/domain/project";
 
 export const projectStoreState: ProjectStoreState = {
   savedLastCommandUnixMillisec: null,
-};
-
-const validateTalkProject = (talkProject: LatestProjectType["talk"]) => {
-  if (
-    !talkProject.audioKeys.every(
-      (audioKey) => audioKey in talkProject.audioItems
-    )
-  ) {
-    throw new Error(
-      "Every audioKey in audioKeys should be a key of audioItems"
-    );
-  }
-  if (
-    !talkProject.audioKeys.every(
-      (audioKey) => talkProject.audioItems[audioKey]?.voice != undefined
-    )
-  ) {
-    throw new Error('Every audioItem should have a "voice" attribute.');
-  }
-  if (
-    !talkProject.audioKeys.every(
-      (audioKey) =>
-        talkProject.audioItems[audioKey]?.voice.engineId != undefined
-    )
-  ) {
-    throw new Error('Every voice should have a "engineId" attribute.');
-  }
-  // FIXME: assert engineId is registered
-  if (
-    !talkProject.audioKeys.every(
-      (audioKey) =>
-        talkProject.audioItems[audioKey]?.voice.speakerId != undefined
-    )
-  ) {
-    throw new Error('Every voice should have a "speakerId" attribute.');
-  }
-  if (
-    !talkProject.audioKeys.every(
-      (audioKey) => talkProject.audioItems[audioKey]?.voice.styleId != undefined
-    )
-  ) {
-    throw new Error('Every voice should have a "styleId" attribute.');
-  }
 };
 
 const applyTalkProjectToStore = async (
@@ -204,6 +141,7 @@ export const projectStore = createPartialStore<ProjectStoreTypes>({
         }
 
         const projectFileErrorMsg = `VOICEVOX Project file "${filePath}" is a invalid file.`;
+        // TODO: ↑のエラーメッセージをちゃんとする
 
         let buf: ArrayBuffer;
         try {
@@ -214,238 +152,23 @@ export const projectStore = createPartialStore<ProjectStoreTypes>({
           await context.dispatch("APPEND_RECENTLY_USED_PROJECT", {
             filePath,
           });
+
           const text = new TextDecoder("utf-8").decode(buf).trim();
           const projectData = JSON.parse(text);
 
-          // appVersion Validation check
-          if (
-            !(
-              "appVersion" in projectData &&
-              typeof projectData.appVersion === "string"
-            )
-          ) {
-            throw new Error(
-              projectFileErrorMsg +
-                " The appVersion of the project file should be string"
-            );
-          }
-          const projectAppVersion: string = projectData.appVersion;
-          if (!semver.valid(projectAppVersion)) {
-            throw new Error(
-              projectFileErrorMsg +
-                ` The app version of the project file "${projectAppVersion}" is invalid. The app version should be a string in semver format.`
-            );
-          }
+          const characterInfos =
+            context.getters.USER_ORDERED_CHARACTER_INFOS("talk");
+          if (characterInfos == undefined)
+            throw new Error("characterInfos == undefined");
 
-          const semverSatisfiesOptions: semver.Options = {
-            includePrerelease: true,
-          };
-
-          // Migration
-          const engineId = EngineId("074fc39e-678b-4c13-8916-ffca8d505d1d");
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.4", semverSatisfiesOptions)
-          ) {
-            for (const audioItemsKey in projectData.audioItems) {
-              if ("charactorIndex" in projectData.audioItems[audioItemsKey]) {
-                projectData.audioItems[audioItemsKey].characterIndex =
-                  projectData.audioItems[audioItemsKey].charactorIndex;
-                delete projectData.audioItems[audioItemsKey].charactorIndex;
-              }
+          const parsedProjectData = await migrateProjectFileObject(
+            projectData,
+            {
+              fetchMoraData: (payload) =>
+                context.dispatch("FETCH_MORA_DATA", payload),
+              characterInfos,
             }
-            for (const audioItemsKey in projectData.audioItems) {
-              if (projectData.audioItems[audioItemsKey].query != null) {
-                projectData.audioItems[audioItemsKey].query.volumeScale = 1;
-                projectData.audioItems[
-                  audioItemsKey
-                ].query.prePhonemeLength = 0.1;
-                projectData.audioItems[
-                  audioItemsKey
-                ].query.postPhonemeLength = 0.1;
-                projectData.audioItems[audioItemsKey].query.outputSamplingRate =
-                  DEFAULT_SAMPLING_RATE;
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.5", semverSatisfiesOptions)
-          ) {
-            for (const audioItemsKey in projectData.audioItems) {
-              const audioItem = projectData.audioItems[audioItemsKey];
-              if (audioItem.query != null) {
-                audioItem.query.outputStereo = false;
-                for (const accentPhrase of audioItem.query.accentPhrases) {
-                  if (accentPhrase.pauseMora) {
-                    accentPhrase.pauseMora.vowelLength = 0;
-                  }
-                  for (const mora of accentPhrase.moras) {
-                    if (mora.consonant) {
-                      mora.consonantLength = 0;
-                    }
-                    mora.vowelLength = 0;
-                  }
-                }
-
-                // set phoneme length
-                // 0.7 未満のプロジェクトファイルは styleId ではなく characterIndex なので、ここだけ characterIndex とした
-                if (audioItem.characterIndex == undefined)
-                  throw new Error("audioItem.characterIndex === undefined");
-                await context
-                  .dispatch("FETCH_MORA_DATA", {
-                    accentPhrases: audioItem.query.accentPhrases,
-                    engineId,
-                    styleId: audioItem.characterIndex,
-                  })
-                  .then((accentPhrases: AccentPhrase[]) => {
-                    accentPhrases.forEach((newAccentPhrase, i) => {
-                      const oldAccentPhrase = audioItem.query.accentPhrases[i];
-                      if (newAccentPhrase.pauseMora) {
-                        oldAccentPhrase.pauseMora.vowelLength =
-                          newAccentPhrase.pauseMora.vowelLength;
-                      }
-                      newAccentPhrase.moras.forEach((mora, j) => {
-                        if (mora.consonant) {
-                          oldAccentPhrase.moras[j].consonantLength =
-                            mora.consonantLength;
-                        }
-                        oldAccentPhrase.moras[j].vowelLength = mora.vowelLength;
-                      });
-                    });
-                  });
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.7", semverSatisfiesOptions)
-          ) {
-            for (const audioItemsKey in projectData.audioItems) {
-              const audioItem = projectData.audioItems[audioItemsKey];
-              if (audioItem.characterIndex != null) {
-                if (audioItem.characterIndex == 0) {
-                  // 四国めたん 0 -> 四国めたん（あまあま） 0
-                  audioItem.speaker = 0;
-                }
-                if (audioItem.characterIndex == 1) {
-                  // ずんだもん 1 -> ずんだもん（あまあま） 1
-                  audioItem.speaker = 1;
-                }
-                delete audioItem.characterIndex;
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.8", semverSatisfiesOptions)
-          ) {
-            for (const audioItemsKey in projectData.audioItems) {
-              const audioItem = projectData.audioItems[audioItemsKey];
-              if (audioItem.speaker != null) {
-                audioItem.styleId = audioItem.speaker;
-                delete audioItem.speaker;
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.14", semverSatisfiesOptions)
-          ) {
-            for (const audioItemsKey in projectData.audioItems) {
-              const audioItem = projectData.audioItems[audioItemsKey];
-              if (audioItem.engineId == undefined) {
-                audioItem.engineId = engineId;
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(projectAppVersion, "<0.15", semverSatisfiesOptions)
-          ) {
-            const characterInfos =
-              context.getters.USER_ORDERED_CHARACTER_INFOS("talk");
-            if (characterInfos == undefined)
-              throw new Error("characterInfos == undefined");
-            for (const audioItemsKey in projectData.audioItems) {
-              const audioItem = projectData.audioItems[audioItemsKey];
-              if (audioItem.voice == undefined) {
-                const oldEngineId = audioItem.engineId;
-                const oldStyleId = audioItem.styleId;
-                const chracterinfo = characterInfos.find((characterInfo) =>
-                  characterInfo.metas.styles.some(
-                    (styeleinfo) =>
-                      styeleinfo.engineId === audioItem.engineId &&
-                      styeleinfo.styleId === audioItem.styleId
-                  )
-                );
-                if (chracterinfo == undefined)
-                  throw new Error(
-                    `chracterinfo == undefined: ${oldEngineId}, ${oldStyleId}`
-                  );
-                const speakerId = chracterinfo.metas.speakerUuid;
-                audioItem.voice = {
-                  engineId: oldEngineId,
-                  speakerId,
-                  styleId: oldStyleId,
-                };
-
-                delete audioItem.engineId;
-                delete audioItem.styleId;
-              }
-            }
-          }
-
-          if (
-            semver.satisfies(
-              projectAppVersion,
-              "<0.16.2",
-              semverSatisfiesOptions
-            )
-          ) {
-            // 0.16.2 未満のプロジェクトファイルはトークの情報のみ
-            // なので全情報(audioKeys/audioItems)をtalkに移動する
-            projectData.talk = {
-              audioKeys: projectData.audioKeys,
-              audioItems: projectData.audioItems,
-            };
-
-            // ソングの情報を初期化
-            // generateSingingStoreInitialScoreが今後変わることがあるかもしれないので、
-            // 0.16.2時点のスコア情報を直接書く
-            projectData.song = {
-              tpqn: DEFAULT_TPQN,
-              tempos: [
-                {
-                  position: 0,
-                  bpm: DEFAULT_BPM,
-                },
-              ],
-              timeSignatures: [
-                {
-                  measureNumber: 1,
-                  beats: DEFAULT_BEATS,
-                  beatType: DEFAULT_BEAT_TYPE,
-                },
-              ],
-              tracks: [
-                {
-                  singer: undefined,
-                  keyRangeAdjustment: 0,
-                  notes: [],
-                },
-              ],
-            };
-
-            delete projectData.audioKeys;
-            delete projectData.audioItems;
-          }
-
-          // Validation check
-          // トークはvalidateTalkProjectで検証する
-          // ソングはSET_SCOREの中の`isValidScore`関数で検証される
-          const parsedProjectData = projectSchema.parse(projectData);
-          validateTalkProject(parsedProjectData.talk);
+          );
 
           if (confirm !== false && context.getters.IS_EDITED) {
             const result = await context.dispatch(
@@ -644,69 +367,3 @@ export const projectStore = createPartialStore<ProjectStoreTypes>({
     },
   },
 });
-
-const moraSchema = z.object({
-  text: z.string(),
-  vowel: z.string(),
-  vowelLength: z.number(),
-  pitch: z.number(),
-  consonant: z.string().optional(),
-  consonantLength: z.number().optional(),
-});
-
-const accentPhraseSchema = z.object({
-  moras: z.array(moraSchema),
-  accent: z.number(),
-  pauseMora: moraSchema.optional(),
-  isInterrogative: z.boolean().optional(),
-});
-
-const audioQuerySchema = z.object({
-  accentPhrases: z.array(accentPhraseSchema),
-  speedScale: z.number(),
-  pitchScale: z.number(),
-  intonationScale: z.number(),
-  volumeScale: z.number(),
-  prePhonemeLength: z.number(),
-  postPhonemeLength: z.number(),
-  outputSamplingRate: z.union([z.number(), z.literal("engineDefault")]),
-  outputStereo: z.boolean(),
-  kana: z.string().optional(),
-});
-
-const morphingInfoSchema = z.object({
-  rate: z.number(),
-  targetEngineId: engineIdSchema,
-  targetSpeakerId: speakerIdSchema,
-  targetStyleId: styleIdSchema,
-});
-
-const audioItemSchema = z.object({
-  text: z.string(),
-  voice: z.object({
-    engineId: engineIdSchema,
-    speakerId: speakerIdSchema,
-    styleId: styleIdSchema,
-  }),
-  query: audioQuerySchema.optional(),
-  presetKey: presetKeySchema.optional(),
-  morphingInfo: morphingInfoSchema.optional(),
-});
-
-const projectSchema = z.object({
-  appVersion: z.string(),
-  talk: z.object({
-    // description: "Attribute keys of audioItems.",
-    audioKeys: z.array(audioKeySchema),
-    // description: "VOICEVOX states per cell",
-    audioItems: z.record(audioKeySchema, audioItemSchema),
-  }),
-  song: z.object({
-    tpqn: z.number(),
-    tempos: z.array(tempoSchema),
-    timeSignatures: z.array(timeSignatureSchema),
-    tracks: z.array(trackSchema),
-  }),
-});
-
-export type LatestProjectType = z.infer<typeof projectSchema>;
